# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Learning Rest API DOCUMENTATION",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dashboard/users/change-password": {
            "put": {
                "tags": [
                    "Dashboard Api Users"
                ],
                "summary": "change password user",
                "operationId": "0efad2e6fa0cab3195e4278c7329612d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "old_password": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string",
                                        "example": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/dashboard/users/upload-image": {
            "post": {
                "tags": [
                    "Dashboard Api Users"
                ],
                "summary": "upload file User",
                "operationId": "4a4689dd18a8a9445e1fd0e71ab43a7a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "type": "file",
                                        "example": "path image"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/courses": {
            "get": {
                "tags": [
                    "Front Api Courses"
                ],
                "summary": "get all courses",
                "operationId": "4d2b332408486c69d28f1200eed63bc8",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "filter courses with user",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "level_id",
                        "in": "query",
                        "description": "filter courses with level",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "filter courses with category",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "semester",
                        "in": "query",
                        "description": "filter courses with semester (first semester , second semester , full semester)",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "filter search name or description courses",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/courses/{course_id}": {
            "get": {
                "tags": [
                    "Front Api Courses"
                ],
                "summary": "show course",
                "operationId": "b5b271252699d2d16e54cf845a7847b6",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/courses/{course_id}/lessons": {
            "get": {
                "tags": [
                    "Front Api Lessons"
                ],
                "summary": "get all lessons",
                "operationId": "b253c99d55efba21db724353b865f900",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "description": "filter lessons with course",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "filter search name or description",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/lessons/{lesson_id}": {
            "get": {
                "tags": [
                    "Front Api Lessons"
                ],
                "summary": "show lesson",
                "operationId": "98ac784bf520d416c74b2cc5019d45cb",
                "parameters": [
                    {
                        "name": "lesson_id",
                        "in": "path",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}